Split main.py into per class files
