+++ kcfetcher/store/__init__.py
@@ -1,0 +1,5 @@
+from .store import Store
+
+__all__ = [
+    Store,
+]
